+++ template/scicloj/ml/generate_api.clj
@@ -1,20 +1,20 @@
 (ns scicloj.ml.generate-api
-  (:require scicloj.ml.metamorph
-            scicloj.ml.core
-            scicloj.ml.dataset
+  (:require scicloj.ml.template.metamorph
+            scicloj.ml.template.core
+            scicloj.ml.template.dataset
             [tech.v3.datatype.export-symbols :as exporter]))
 
 (defn generate-api! [_]
 
-  (exporter/write-api! 'scicloj.ml.metamorph
+  (exporter/write-api! 'scicloj.ml.template.metamorph
                        'scicloj.ml.metamorph
                        "src/scicloj/ml/metamorph.clj"
                        [])
-  (exporter/write-api! 'scicloj.ml.dataset
+  (exporter/write-api! 'scicloj.ml.template.dataset
                        'scicloj.ml.dataset
                        "src/scicloj/ml/dataset.clj"
                        [])
-  (exporter/write-api! 'scicloj.ml.core
+  (exporter/write-api! 'scicloj.ml.template.core
                        'scicloj.ml.core
                        "src/scicloj/ml/core.clj"
                        []))